refactor: Game 도메인 -> 컨트롤러
- 하드 코딩 상수화
